requirements.txt
langgraph==1.1.1
langchain-aws==1.4.0
boto3==1.42.65
streamlit==1.42.0
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.0
reportlab==4.2.0
polars==1.21.0
langchain-ollama==1.0.1   